Apply first round of changes from PR comments.
