+++ Resources/python/pyproject.toml
@@ -92,15 +92,15 @@ minari = [
 
 # Dependencies for RLlib offline algorithms (BC, MARWIL).
 # RLlib reads offline episodes as msgpack-serialized Parquet.
-offline = [
+rllib-offline = [
     "schola[rllib]",
     "msgpack",
     "msgpack-numpy",
 ]
 
-# All dependencies (sb3, rllib, minari, offline).
+# All dependencies (sb3, rllib, minari, rllib-offline).
 all = [
-    "schola[sb3,rllib,minari,offline]",
+    "schola[sb3,rllib,minari,rllib-offline]",
 ]
 
 [project.scripts]
