+++ Resources/python/schola/scripts/rllib/collect/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026 Advanced Micro Devices, Inc. All Rights Reserved.
+
+"""Run RLlib demonstration collection as a module."""
+
+from .collect import app as app
+
+if __name__ == "__main__":
+    app.meta()
