+++ Resources/python/schola/scripts/rllib/offline_train/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026 Advanced Micro Devices, Inc. All Rights Reserved.
+
+"""Run RLlib offline training as a module."""
+
+from .offline_train import app as app
+
+if __name__ == "__main__":
+    app.meta()
